+++ practice/lab4/shapes/solidshapes/circle.py
@@ -10,7 +10,7 @@ class Circle(SolidShape):
     _outline_color: int
     _fill_color: int
 
-    def __init__(self, x: int, y: int, radius: float, outline_color: int, fill_color: int):
+    def __init__(self, x: float, y: float, radius: float, outline_color: int, fill_color: int):
         self._center = Point(x, y)
         self._radius = radius
         self._outline_color = outline_color
